Results Found
This is a folder of all the graphs and statistics found when running our R code. We did not include all of them in our final paper, but I decided to include them all here in case it helps with marking. Also, it makes it easier to collaborate for making the presentation based on everything found.