R code used for the Project
This is the R file I constructed in analyzing the 10,000 most popular English words. This File when run in the same directory as "google-10000-english-no-swears.txt" produces our total results. Our total results, as seen in the folder for this repository named "Results Found" is an example of what one would see when this R file is run properly (with respect to location of the txt file).